requirements.txt
requests
pydantic
pymongo
python-dotenv